size_t in loops

+++ solution/src/transform.c
@@ -14,8 +14,8 @@ struct image none(const struct image* input_image) {
 
 struct image rotate_90_ccw(const struct image* input_image) {
     struct image transformed = create_image(input_image->height, input_image->width);
-    for (uint64_t y = 0; y < input_image->height; y++) {
-        for (uint64_t x = 0; x < input_image->width; x++) {
+    for (size_t y = 0; y < input_image->height; y++) {
+        for (size_t x = 0; x < input_image->width; x++) {
             transformed.data[x * transformed.width + (input_image->height - y - 1)] = input_image->data[y * input_image->width + x];
         }
     }
@@ -24,8 +24,8 @@ struct image rotate_90_ccw(const struct image* input_image) {
 
 struct image rotate_90_cw(const struct image* input_image) {
     struct image transformed = create_image(input_image->height, input_image->width);
-    for (uint64_t y = 0; y < input_image->height; y++) {
-        for (uint64_t x = 0; x < input_image->width; x++) {
+    for (size_t y = 0; y < input_image->height; y++) {
+        for (size_t x = 0; x < input_image->width; x++) {
             transformed.data[(input_image->width - x - 1) * transformed.width + y] = input_image->data[y * input_image->width + x];
         }
     }
@@ -34,8 +34,8 @@ struct image rotate_90_cw(const struct image* input_image) {
 
 struct image flip_horizontal(const struct image* input_image) {
     struct image transformed = create_image(input_image->width, input_image->height);
-    for (uint64_t y = 0; y < input_image->height; y++) {
-        for (uint64_t x = 0; x < input_image->width; x++) {
+    for (size_t y = 0; y < input_image->height; y++) {
+        for (size_t x = 0; x < input_image->width; x++) {
             transformed.data[y * transformed.width + (input_image->width - x - 1)] = input_image->data[y * input_image->width + x];
         }
     }
@@ -44,8 +44,8 @@ struct image flip_horizontal(const struct image* input_image) {
 
 struct image flip_vertical(const struct image* input_image) {
     struct image transformed = create_image(input_image->width, input_image->height);
-    for (uint64_t y = 0; y < input_image->height; y++) {
-        for (uint64_t x = 0; x < input_image->width; x++) {
+    for (size_t y = 0; y < input_image->height; y++) {
+        for (size_t x = 0; x < input_image->width; x++) {
             transformed.data[(input_image->height - y - 1) * transformed.width + x] = input_image->data[y * input_image->width + x];
         }
     }
